BBS/requirements.txt
alembic
amqp
anyjson
billiard
blinker
celery
certifi
chardet
click
eventlet
Flask
Flask-Mail
Flask-Migrate
flask-paginate
Flask-Script
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
itsdangerous
Jinja2
kombu
Mako
MarkupSafe
Pillow
PyMySQL
python-dateutil
python-editor
python-memcached
pytz
qiniu
redis
requests
shortuuid
six
SQLAlchemy
urllib3
vine
Werkzeug
WTForms
